backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
sqlmodel==0.0.22
requests==2.32.3
python-dotenv==1.0.1
psycopg[binary]==3.3.2
